// projects/ccf-database/src/lib/interfaces.d.ts
export interface SpatialSearch {
    /** X coordinate relative to target in millimeters */
    x: number;
    /** y coordinate relative to target in millimeters */
    y: number;
    /** z coordinate relative to target in millimeters */
    z: number;
    /** Size of the probing sphere in millimeters */
    radius: number;
    /** The target spatial entity IRI */
    target: string;
}